Update FSRTracker.toc
hard locks the 5 second timer in first. So nothing can falsely trigger it while still inside the 5 second window and only tries to let the Mana you Regen by the Five Second Rule through, and tries to ignore smaller gains or bigger spikes etc...

But still let's Evocate through since it ticks with your Mana Regen and it's nice to have a visual on when the last tick will be so you can end the cast early...

Drinking is also allowed to bypass the logic that checks for mana spikes so it doesnt stop prevent or stop the 2s pulses...

Innervate I've also let through, the tooltip makes it sound like you just keep your normal mana gain while casting. So I've implemented something where it will detect if Innervate is in our buffs and if so stop all FSR spell triggers from coming in and sync FSR to Innervate on the first mana tick after the buff is detected, then after innervate falls off wait for another mana tick or new trigger spell cast to resync the timer or mana full to end it...

Beware that I have not been able to test the addon's behavior with Innervate. Don't have a druid at my disposal to test with, so let me know if any issues pop up regarding Innervate, or anything else for that matter ! ^^